pyproject.toml
[tool.poetry]
name = "python-osc-test"
version = "0.1.0"
description = ""
authors = ["Ri <45131445+rl2999@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.7"
python-osc = "^1.7.4"
pyserial = "^3.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
